// types/global.d.ts
export type UserData = {
  id:             string
  name?:          string
  registered?:    boolean
  regName?:       string
  regAge?:        string
  regDate?:       string
  level?:         number
  exp?:           number
  premium?:       boolean
  premiumExpiry?: number | null
  banned?:        boolean
  [key: string]:  unknown
}

export type GroupData = {
  id:            string
  subject?:      string
  selfmode?:     boolean
  antilink?:     boolean
  antigrup?:     boolean
  antilinkmode?: string
  welcome?:      boolean
  welcomemsg?:   string
  goodbye?:      boolean
  goodbyemsg?:   string
  openclose?:    boolean
  nsfw?:         boolean
  intro?:        boolean
  [key: string]: unknown
}

export type LidMapData = Record<string, string>